Merge pull request #11
LustratedPersonsRegister integrated
Integrated Unified State Register of Lustrated Persons. Docstrings added. Refactoring and optimization

+++ src/DebtorsRegister.py
@@ -1,17 +1,17 @@
 import gc
 import json
 import logging
+import mmap
 import os
 import shutil
 import zipfile
-import mmap
 from datetime import datetime
 from io import BytesIO
-from pymongo.errors import PyMongoError
 
 import requests
 from dask import dataframe as dd
 from prettytable import PrettyTable
+from pymongo.errors import PyMongoError
 
 from src.dataset import Dataset
 
@@ -27,7 +27,7 @@ class DebtorsRegister(Dataset):
             general_dataset = requests.get(
                 'https://data.gov.ua/api/3/action/package_show?id=506734bf-2480-448c-a2b4-90b6d06df11e').text
         except ConnectionError:
-            logging.error('Error during general DebtorsRegister dataset JSON receiving occured')
+            logging.error('Error during general DebtorsRegister dataset JSON receiving occurred')
             print('Error during dataset receiving occurred!')
         else:
             general_dataset_json = json.loads(general_dataset)
@@ -39,7 +39,7 @@ class DebtorsRegister(Dataset):
             debtors_general_dataset_id_json = requests.get(
                 'https://data.gov.ua/api/3/action/resource_show?id=' + debtors_general_dataset_id).text
         except ConnectionError:
-            logging.error('Error during DebtorsRegisterr resources JSON id receiving occured')
+            logging.error('Error during DebtorsRegister resources JSON id receiving occurred')
             print('Error during dataset receiving occurred!')
         else:
             debtors_general_dataset_json = json.loads(debtors_general_dataset_id_json)
@@ -55,8 +55,8 @@ class DebtorsRegister(Dataset):
             # get ZIP file
             debtors_dataset_zip = requests.get(zip_url).content
         except OSError:
-            logging.error('Error during DebtorsRegisterZIP receiving occured')
-            print('Error during ZIP file receiving occured!')
+            logging.error('Error during DebtorsRegisterZIP receiving occurred')
+            print('Error during ZIP file receiving occurred!')
         else:
             logging.info('A DebtorsRegister dataset received')
             # get lists of file
@@ -101,10 +101,11 @@ class DebtorsRegister(Dataset):
 
     @Dataset.measure_execution_time
     def __clear_collection(self):
-        debtors_col = self.db['Debtors']
-        count_deleted_documents = debtors_col.delete_many({})
-        logging.warning('%s documents deleted. The wanted persons collection is empty.', str(
-            count_deleted_documents.deleted_count))
+        if self.is_collection_exists('Debtors'):
+            debtors_col = self.db['Debtors']
+            count_deleted_documents = debtors_col.delete_many({})
+            logging.warning('%s documents deleted. The wanted persons collection is empty.', str(
+                count_deleted_documents.deleted_count))
 
     @Dataset.measure_execution_time
     def __create_service_json(self):
@@ -136,7 +137,8 @@ class DebtorsRegister(Dataset):
     def __update_metadata(self):
         collections_list = self.db.list_collection_names()
         # update or create DebtorsRegisterServiceJson
-        if ('ServiceCollection' in collections_list) and (self.serviceCol.count_documents({'_id': 3}, limit=1) != 0):
+        if (self.is_collection_exists('ServiceCollection')) and (
+                self.serviceCol.count_documents({'_id': 3}, limit=1) != 0):
             self.__update_service_json()
             logging.info('DebtorsRegisterServiceJson updated')
         else:
@@ -145,10 +147,11 @@ class DebtorsRegister(Dataset):
 
     @Dataset.measure_execution_time
     def __delete_collection_index(self):
-        debtors_col = self.db['Debtors']
-        if 'full_text' in debtors_col.index_information():
-            debtors_col.drop_index('full_text')
-            logging.warning('Debtors Text index deleted')
+        if self.is_collection_exists('Debtors'):
+            debtors_col = self.db['Debtors']
+            if 'full_text' in debtors_col.index_information():
+                debtors_col.drop_index('full_text')
+                logging.warning('Debtors Text index deleted')
 
     @Dataset.measure_execution_time
     def __create_collection_index(self):
@@ -178,17 +181,17 @@ class DebtorsRegister(Dataset):
                 result_table._max_width = {'DEBTOR NAME': 25, 'PUBLISHER': 25, 'EXECUTIVE SERVICE': 35,
                                            'EXECUTIVE SERVICE EMPLOYEE': 25, 'CATEGORY': 25}
                 # show only 10 first search results
-                for result in debtors_col.find({'$text': {'$search': query_string}}, {'score': {'$meta': 'textScore'}})\
+                for result in debtors_col.find({'$text': {'$search': query_string}}, {'score': {'$meta': 'textScore'}}) \
                         .sort([('score', {'$meta': 'textScore'})]).limit(10).allow_disk_use(True):
                     result_table.add_row([result['DEBTOR_NAME'], result['DEBTOR_CODE'], result['PUBLISHER'],
                                           result['EMP_ORG'], result['EMP_FULL_FIO'], result['VD_CAT']])
                 print(result_table.get_string(
-                    title='The debtors register: ' + str(result_count) + ' records found'))
+                    title=f'The debtors register: {result_count} records found'))
                 logging.warning(
                     'The debtors register: %s records found', str(result_count))
                 print('Only 10 first search results showed')
                 # save all search results into HTML
-                for result in debtors_col.find({'$text': {'$search': query_string}}, {'score': {'$meta': 'textScore'}}).\
+                for result in debtors_col.find({'$text': {'$search': query_string}}, {'score': {'$meta': 'textScore'}}). \
                         sort([('score', {'$meta': 'textScore'})]).allow_disk_use(True):
                     result_table.add_row([result['DEBTOR_NAME'], result['DEBTOR_CODE'], result['PUBLISHER'],
                                           result['EMP_ORG'], result['EMP_FULL_FIO'], result['VD_CAT']])
@@ -205,7 +208,7 @@ class DebtorsRegister(Dataset):
     def setup_dataset(self):
         self.__delete_collection_index()
         self.__clear_collection()
-        __debtorsDatasetZIPUrl = self.__get_dataset()
-        self.__save_dataset(__debtorsDatasetZIPUrl)
+        __debtors_dataset_zip_url = self.__get_dataset()
+        self.__save_dataset(__debtors_dataset_zip_url)
         self.__update_metadata()
         self.__create_collection_index()
